requirements.txt
boto3==1.16.59
botocore==1.19.59
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
s3transfer==0.3.4
schedule==0.6.0
six==1.15.0
urllib3==1.26.3
